feat(storybook): setup LOCALE_ID with browser language
